requirements.txt
alembic==1.7.5
aniso8601==9.0.1
black==21.12b0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
mypy-extensions==0.4.3
password-strength==0.0.3.post2
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
tomli==1.2.3
typing-extensions==4.0.1
Werkzeug==2.0.2
